recipes and recipe screen and vms

+++ lib/Models/Ingredient.dart
@@ -3,12 +3,12 @@ import 'package:personal_recipes/Enums/enums.dart';
 class Ingredient {
   String title;
   String unit;
-  int amount;
+  double amount;
 
   Ingredient({required this.title, required this.unit, required this.amount});
 
   Ingredient.fromJSON(Map<String, dynamic> ingredient)
       : title = ingredient['title'],
         unit = ingredient['unit'],
-        amount = ingredient['amount'];
+        amount = ingredient['amount'].toDouble();
 }
